server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
opencv-python
Pillow
timm
einops
numpy